comments and remove compilation warnings

+++ libblankenhain/src/ExampleInstrument.cpp
@@ -36,7 +36,7 @@ void ExampleInstrument::processVoice(VoiceState& voice, unsigned int timeInSampl
 	float release = getInterpolatedParameter(7).get();
 	unsigned int mode = static_cast<unsigned int>(getInterpolatedParameter(8).get());
 	
-	this->osc.setFrequency(aux::noteToFrequency(voice.key));
+	this->osc.setFrequency(aux::noteToFrequency(static_cast<float>(voice.key)));
 
 	if (mode == 0u)
 		currentSound = &osc;
